Don't set boot_disk in preseed
When a new source of debconf answers is provided, they're added to the
the debconf cache such that they override any already-answered
questions. Since the preseed is loaded after the settings passed on the
kernel command line, the preseed takes precedence. However, we want to
be able to specify the install disk on a per-node basis (as this
differs a little bit across systems), so this needs to be passed on the
kernel command line and not be set in the preseed.