dhcp and dns templates should call cobbler sync
I found that the dhcp server was not started on a freshly installed
system. More exploration showed that the /etc/dhcp.conf file also did
not exist.

After reading some docs[1][2], I saw that running 'cobbler sync' is required
for the dhcp template to populate the config file correctly.

I experimented with changing the cobbler service to call cobbler sync
instead of restart, but found the restart call is also required. I also
found through testing that the cobbler sync call must be called after the
cobbler service is started.

My one concern here is that others have reported that the current configuration
works. I do not believe my findings are consistent with this, so it makes me thing
they may be something specific to my environment.

[1] https://github.com/cobbler/cobbler/wiki/DHCP%20management
[2] https://github.com/cobbler/cobbler/wiki/DNS%20management