Move cobblersystem management to native type
This fix was initially motivated by:
  https://bugs.launchpad.net/openstack-cisco/+bug/1190319

It removes the exec that was being used to manage
cobblersystems and replaces it with a native type.

The actual type code was based on code from the internet:

  http://forge.puppetlabs.com/jsosic/cobbler

It also required a non-trivial amount of patches
to ensure that is supports all of the options that we
need.