Parameterize Cisco package repo setup
Previous versions of this module relied on static strings in the
late_command portion of the preseed template in order to set up
the Cisco package repositories in the preseed file passed to nodes
being PXE booted.  This meant that with every release, the
release manager would need to push a change to the file to set up
the repository properly for that release.  It also meant that
repository information specified for the puppet-coe module could
wind up being inconsistent with the information in the preseed,
thus generally violated DRY principles.  This patch parameterizes
the apt repo setup portion of the preseed.

Partial-Bug: #1269856