Remove unnecessary agent call in late_command
In late_command we call the Puppet agent.  This is theoretically to
ensure that we get a clean PKI setup: the agent sets up a CSR and
requests a signature on it shortly after we ran ntpdate, so we can
be reasonably sure that we won't have PKI problems due to the
master seeing a CSR with a future datestamp on it due to clocks
being out of sync.

However, we no longer need this: we also set up ntpd as part of the
preseed process, so clocks get synced on boot anyway (even if the
ntpdate sync doesn't survive for some reason). Further, debugging
PKI issues is easier if the CSR generation/signing are done on first
boot anyway since you have a full environment vs a chrooted one.
We're also getting reports that occasionally in some circumstances
the agent run in preseed causes delays and other issues, so given
that we don't need it anyway we should just remove it.

Closes-Bug: #1282135